pyproject.toml
[tool.poetry]
name = "ml-image-generator"
version = "0.1.0"
description = ""
authors = ["Rafał Majewski <goodheropl@gmail.com>"]
license = "CC0-1.0"

[tool.poetry.dependencies]
python = "^3.10"
torch = "^1.11.0"
Pillow = "^9.1.1"
numpy = "^1.22.4"

[tool.poetry.dev-dependencies]
mypy = "^0.960"
flake8 = "^4.0.1"
tox = "^3.25.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
